# Define user model attributes